+++ Changelog.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.0.1] - 2020-08-01
+## [1.0.2] - 2024-11-15
+
+### Fixed
+- Fixed an issue where the single record query would not work
+- Fixed an issue where the query for records without language field would not work
+
+## [1.0.1] - 2024-08-01
 
 ### Fixed
 - Fixed issue where filters were not being applied to the query